pyproject.toml
[tool.poetry]
name = "new-year-2024"
version = "0.1.0"
description = ""
authors = ["takikawauniv <takiawauniv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytk = "^0.0.2.1"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
